Fix unit Tests

+++ x10/gen/CodeGenerator.cs
@@ -46,7 +46,8 @@ namespace x10.gen {
       AllUiDefinitions = allUiDefinitions;
       _platformLibraries = platformLibraries;
 
-      Directory.Delete(rootGenerateDir, true);
+      if (Directory.Exists(rootGenerateDir))
+        Directory.Delete(rootGenerateDir, true);
 
       foreach (Entity entity in AllEntities.All.Where(x => !x.IsAbstract))
         Generate(entity);
